WSD ID 7962, 9104, 11191, 11254 and 7969. Pull request for the quarterly updates.  (#114)
* Completed the first draft of the APM spring boot on K8S workshop for OCI

Completed the first draft of the APM spring boot on K8S workshop for OCI

* Updating task titles etc

Updating task titles etc

* Updated the estimated time

Updated the estimated time after measuring the time with the self QA

* Changed lab directory names

Split the first lab which resulted in directory names as numbering changed. also added minor fixes.

* Numerous updates done to the APM SB-HOL

Self QA completed

* minor adjustments to go with the self-qa task list

* corrected a help link and a typo

* Corrected the format to go with the guidelines

* added changes based on the team QA feedbacks

* additional changes based on the feedbacks

* corrected a typo

* Additional fix based on the feedbacks

* first draft of the WLS lab for OCW

* Corrected a typo in a service name configured for APM java agent install command

* first submission of the WSM 11254

* applied enhancements to the APM demo lab

* a few minor fixes done based on the lint checker

* title fix to comply the lint checker

* spelling corrections

* updated the estimated time

* Updated the APM with WLS HOL

* Updated based on the feedbacks and spelling checker

* APM Helidon lab - updates for ocw

Helidon lab - updating the folder structure, and made changes to go with the latest LL guidelines

* Self QA on helidon apm lab.

added changes based on the self QA

* additional updates to the APM helidon workshop

* helidon-apm lab -self qa. additional changes

* Update 3-1-1-2-filesystem.png

* Updated WLS APM workshop with Lint checker result

* Update access-env.md

* Updated the Help section for all 4 hols

added ocw folders in the workshops directory that are specific to OCW

* Updated Index.html for 4 APM workshops

* Editing made based on the non-team feedbacks.

* Update create-file-system.md

corrected a typo

* Update introduction.md

corrected the estimated lab time

* Update introduction.md

corrected the estimated lab time

* corrected missing & misplaced steps

* adjustments required as the Java agent version has updated

* Added steps to the workshop as the Green button was enabled.

Added steps to the workshop as the Green button was enabled. Also some minor updates.

* Updated the clean up steps including the deletion of domain and compartment

* Updated the acknowledgement section

Updated Avi's new title, added Anand for a contributor.

* removed steps to enter name at the compartment deletion

* minor spacing adjustments on names at acknowledgements

* Update cleanup.md

Removed extra spaces that resulted in failing to display an image

* Added a promotion to another HOL at the end of the workshop

* Updated a note section, mentioning a new feature

Updated a notes section to add mention to a new feature introduced recently

* added additional image to the workshop intro section

* adding introduction lab to two other hols.

added a lab that introduces get started HOL, to APM workshops.

* Made the promotion labs optional

* running periodic maintenance for Native cloud app APM workshop

adding numbers in folder names

* updated links to other APM workshops

* Quarterly update on springboot workshop

Updated lab steps to go with the Java agent 1.8 and  OCI UI changes

* Updated the WLS workshop with Agent 1.8 & general OCI UI changes.

* Revert "running periodic maintenance for Native cloud app APM workshop"

This reverts commit 448158bb3bbf82b6f99cfbac4df0d9f5706da3d2.

* Updated image file names

Updated image file names so that the cloudshell menu item image and cloushell console images can be identified from the names

* Replaced the image with the corporate approved one

Replaced the image file for the cloud-shell menu item with the suggested image.

* Replaced the local file image with the file link from the corporate page

Correction to my previous change. Removed the local file and added a link to the shared location for a common image.